Cluster on Overlay Quickstart
Added a shortcut command for running a node on an overlay network.
The overlay address for the container is resolved automatically, and
join addresses are resolved by container hostname rather than
predetermined IP address.

Running in this mode is perfect for running ad-hoc service discovery
within a stack. This eliminates application (reverse proxy) dependency
on SD infrastructure running below the container layer.

Signed-off-by: Jeff Nickoloff <jeff@allingeek.com>